requirements.txt
aiofiles==24.1.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dlib==19.24.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.0
fonttools==4.55.3
geojson==3.1.0
greenlet==3.1.1
h11==0.14.0
httptools==0.6.1
idna==3.10
ignr==2.2
imutils==0.5.4
jwt==1.3.1
kiwisolver==1.4.7
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
minio==7.2.15
numpy==2.1.1
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
packaging==24.2
passlib==1.7.4
pillow==11.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pyparsing==3.2.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.11
pytz==2024.2
PyYAML==6.0.2
rsa==4.9
setuptools==75.8.0
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.31.0
uvloop==0.20.0
vulture==2.13
watchfiles==0.24.0
websockets==13.1
